app to src rename config updates to match current domain
