+++ backend/.env
@@ -3,6 +3,6 @@ POSTGRES_PASSWORD=postgres
 POSTGRES_USER=postgres
 POSTGRES_DB=pix
 DATABASE_PORT=5432
-POSTGRES_HOST=localhost
+POSTGRES_HOST=pix.cloud.ut.ee
 
-CLIENT_ORIGIN=http://localhost:5173
+CLIENT_ORIGIN=http://pix.cloud.ut.ee:5173
