+++ backend/Dockerfile
@@ -16,6 +16,9 @@ RUN poetry config virtualenvs.create false && poetry install --no-dev
 # Copy the rest of the application code to the container
 COPY . .
 
+# Run the init.py script
+RUN python database/init.py
+
 # Expose port 8000 for the application
 EXPOSE 8000
