+++ backend/main.py
@@ -2,7 +2,7 @@ from fastapi import FastAPI, UploadFile, File, Form
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 
-from app.routers import project, file, user
+from src.routers import project, file, user
 
 app = FastAPI()
