+++ nginx.conf
@@ -11,20 +11,23 @@ http {
   sendfile on;
   keepalive_timeout 65;
 
-
-
   server {
     listen 80;
     server_name pix.cloud.ut.ee;
+
     root   /usr/share/nginx/html;
     index  index.html index.htm;
 
     location / {
-      try_files $uri $uri/ /index.html;
+      proxy_pass http://client:80;
+      proxy_set_header Host $http_host;
+      proxy_set_header X-Real-IP $remote_addr;
+      proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+      proxy_set_header X-Forwarded-Proto $scheme;
     }
 
-    location /api {
-      proxy_pass http://api:8000;
+    location /api/ {
+      proxy_pass http://api:8000/api/;
       proxy_set_header Host $http_host;
       proxy_set_header X-Real-IP $remote_addr;
       proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
